requirements.txt
git+https://github.com/openai/swarm.git
python-dotenv
